+++ src/lib/events.test.ts
@@ -1,7 +1,11 @@
 import {
+  ApiKeyPlacement,
   AstMetadata,
+  HttpScheme,
   MapDocumentNode,
   ProfileDocumentNode,
+  ProviderJson,
+  SecurityType,
 } from '@superfaceai/ast';
 import { getLocal } from 'mockttp';
 
@@ -129,6 +133,36 @@ const mockMapDocument: MapDocumentNode = {
   ],
 };
 
+const mockProviderJson = (name: string): ProviderJson => ({
+  name,
+  services: [{ id: 'test-service', baseUrl: 'service/base/url' }],
+  securitySchemes: [
+    {
+      type: SecurityType.HTTP,
+      id: 'basic',
+      scheme: HttpScheme.BASIC,
+    },
+    {
+      id: 'api',
+      type: SecurityType.APIKEY,
+      in: ApiKeyPlacement.HEADER,
+      name: 'Authorization',
+    },
+    {
+      id: 'bearer',
+      type: SecurityType.HTTP,
+      scheme: HttpScheme.BEARER,
+      bearerFormat: 'some',
+    },
+    {
+      id: 'digest',
+      type: SecurityType.HTTP,
+      scheme: HttpScheme.DIGEST,
+    },
+  ],
+  defaultService: 'test-service',
+});
+
 const mockServer = getLocal();
 
 describe('events', () => {
@@ -147,7 +181,7 @@ describe('events', () => {
     const profile = new BoundProfileProvider(
       mockProfileDocument,
       mockMapDocument,
-      'provider',
+      mockProviderJson('provider'),
       {
         services: ServiceSelector.withDefaultUrl(mockServer.url),
         security: [],
@@ -187,7 +221,7 @@ describe('events', () => {
     const profile = new BoundProfileProvider(
       mockProfileDocument,
       mockMapDocument,
-      'someprovider',
+      mockProviderJson('someprovider'),
       {
         services: ServiceSelector.withDefaultUrl(
           'https://unreachable.localhost'
@@ -225,7 +259,7 @@ describe('events', () => {
     const profile = new BoundProfileProvider(
       mockProfileDocument,
       mockMapDocument,
-      'provider',
+      mockProviderJson('provider'),
       {
         services: ServiceSelector.withDefaultUrl(mockServer.url),
         security: [],
@@ -255,7 +289,7 @@ describe('events', () => {
     const profile = new BoundProfileProvider(
       mockProfileDocument,
       mockMapDocument,
-      'provider',
+      mockProviderJson('provider'),
       {
         services: ServiceSelector.withDefaultUrl(mockServer.url),
         security: [],
@@ -285,7 +319,7 @@ describe('events', () => {
     const profile = new BoundProfileProvider(
       mockProfileDocument,
       mockMapDocument,
-      'provider',
+      mockProviderJson('provider'),
       {
         services: ServiceSelector.withDefaultUrl(mockServer.url),
         security: [],