+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.5.1] - 2022-06-14
 ### Fixed
 - Do not recache ProviderConfiguration on security values change
+- Do not call `Config.instance()` in the global scope to defer reading of env variables
+- Refactor security values passing in `perform`
 
 ## [1.5.0] - 2022-06-09
 ### Added